Changed directory structure to include cpp-markdown files in separate directory
